+++ src/Control/Effect/Accum/Internal.hs
@@ -1,0 +1,12 @@
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE KindSignatures #-}
+module Control.Effect.Accum.Internal
+( Accum(..)
+) where
+
+import Data.Kind (Type)
+
+-- | @since 1.1.2.0
+data Accum w (m :: Type -> Type) k where
+  Add  :: w -> Accum w m ()
+  Look ::      Accum w m w