+++ ChangeLog.md
@@ -1,8 +1,19 @@
+# v1.1.2.0
+
+- Adds `Accum` ([#391](https://github.com/fused-effects/fused-effects/pull/391)) (by @turion)
+  - Adds an `Accum` effect
+  - Adds a church-encoded `Control.Carrier.Accum.Church` carrier
+  - Adds a strict `Control.Carrier.Accum.Church` carrier
+  - Defines `Algebra` instances for the two mentioned carriers,
+    and for `Control.Monad.Trans.Accum` from `transformers`
+
 - Defines, `Algebra`, `Alternative`, `Applicative`, `Foldable`, `Functor`, `Monad`, `MonadFail`, `MonadFix`, `MonadIO`, `MonadPlus`, `MonadTrans`, `MonadZip`, and `Traversable` instances for `Control.Effect.Choose.Choosing`. ([#419](https://github.com/fused-effects/fused-effects/pull/419))
+
 
 # v1.1.1.2
 
 - Adds support for `ghc` 9.2.1 and `base` 4.16.
+
 
 # v1.1.1.1